+++ packages/jet/common/config.ts
@@ -16,6 +16,7 @@ export const DefaultConfig = {
   outDir: DefaultOutDir,
   dev: {
     stage: undefined as string | undefined,
+    stacks: undefined as string[] | undefined,
     watcher: {
       watch: ['lib/**/*'],
       ignore: ['node_modules'],
@@ -25,6 +26,7 @@ export const DefaultConfig = {
   },
   deploy: {
     stage: undefined as string | undefined,
+    stacks: undefined as string[] | undefined,
     deployArgs: [] as string[],
   },
   projectDir: '.',