+++ packages/jet/flight/core/args.ts
@@ -6,6 +6,7 @@ export type Command = 'dev' | 'deploy' | 'list-stages';
 export type Args = {
   command?: Command;
   stage?: string;
+  stacks?: string[];
   synthArgs?: string[];
   deployArgs?: string[];
   config?: string;
@@ -21,6 +22,10 @@ export async function setupArgs(): Promise<Args> {
           type: 'string',
           description: 'Stage to use for development',
         })
+        .option('stacks', {
+          type: 'array',
+          description: 'Stacks to dev against',
+        })
         .option('synth-args', {
           type: 'array',
           description: 'Extra arguments to cdk synth',
@@ -35,6 +40,10 @@ export async function setupArgs(): Promise<Args> {
         .positional('stage', {
           type: 'string',
           description: 'Stage to use for development',
+        })
+        .option('stacks', {
+          type: 'array',
+          description: 'Stacks to deploy',
         })
         .option('deploy-args', {
           type: 'array',
@@ -61,6 +70,7 @@ export async function setupArgs(): Promise<Args> {
   return {
     command: args._[0] as Command,
     stage: args.stage,
+    stacks: args.stacks as string[],
     synthArgs: args['synth-args'] as string[],
     deployArgs: args['deploy-args'] as string[],
     config: args.config,